backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-allauth==65.5.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0